+++ constants.js
@@ -1,0 +1,1 @@
+export const TOKEN_CONTRACT_ADDRESS = "";
